+++ src/Exceptions/BadRequestException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Conlect\ImageIIIF\Exceptions;
+
+class BadRequestException extends BaseException
+{
+    public function __construct(string $message = 'Bad Request', int $code = 400)
+    {
+        parent::__construct($message, $code);
+    }
+}