+++ src/Validators/Contracts/ValidatorInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Conlect\ImageIIIF\Validators\Contracts;
+
+interface ValidatorInterface
+{
+    public function valid($value);
+}