+++ src/Validators/ValidatorAbstract.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Conlect\ImageIIIF\Validators;
+
+abstract class ValidatorAbstract
+{
+    public function __construct($config)
+    {
+        $this->config = $config;
+    }
+}